+++ assetActions.config.js
@@ -1,19 +1,20 @@
-import CopcViewer from './src/actions/assets/CopcViewer.js';
-import GeoJsonIo from './src/actions/assets/GeoJsonIo.js';
 //import Felt from './src/actions/assets/Felt.js';
+import Cesium from './src/actions/assets/Cesium.js';
+import CopcViewer from './src/actions/assets/CopcViewer.js';
+import F3D from './src/actions/assets/F3D.js';
+import GeoJsonIo from './src/actions/assets/GeoJsonIo.js';
+// import Geofox from './src/actions/assets/Geofox.js';
 // import Geoparquet from './src/actions/assets/Geoparquet.js';
 import Potree from './src/actions/assets/Potree.js';
 import Protomaps from './src/actions/assets/Protomaps.js';
-import OGC3dTiles from './src/actions/assets/OGC3dTiles.js';
-import F3D from './src/actions/assets/F3D.js';
 
 export default {
   //Felt
-  OGC3dTiles,
+  Cesium,
   CopcViewer,
+  F3D,
   GeoJsonIo,
   // Geoparquet, // not ready yet
   Potree,
   Protomaps,
-  F3D,
 };
