+++ linkActions.config.js
@@ -1,9 +1,10 @@
 // import Felt from './src/actions/links/Felt.js';
-import OGC3dTiles from './src/actions/links/OGC3dTiles.js';
+// import Geofox from './src/actions/links/Geofox.js';
+import Cesium from './src/actions/links/Cesium.js';
 import Protomaps from './src/actions/links/Protomaps.js';
 
 export default {
     // Felt, 
-    OGC3dTiles, 
+    Cesium, 
     Protomaps
 };
