+++ src/actions/assets/CopcViewer.js
@@ -7,8 +7,7 @@ export default class CopcViewer extends AssetActionPlugin {
   get show() {
     return this.component.isBrowserProtocol && (
       this.asset.type === 'application/vnd.laszip+copc'
-      || this.asset.href.toLowerCase().includes('ept.json')
-
+      ||  URI(this.asset.href).filename() == 'ept.json'
     );
   }
