+++ src/actions/assets/Protomaps.js
@@ -13,7 +13,7 @@ export default class Protomaps extends AssetActionPlugin {
     // Rather check if .pmtiles substring present in this.asset.href or simply this.component.filename.endsWith('pmtiles')
     return this.component.isBrowserProtocol && (
       PROTOMAPS_SUPPORTED_TYPES.includes(this.asset.type)
-      || this.asset.href.toLowerCase().includes('.pmtiles')
+      ||  URI(this.asset.href).suffix() == 'pmtiles'
     );
   }
