+++ src/actions/links/Cesium.js
@@ -1,23 +1,17 @@
-import AssetActionPlugin from "../AssetActionPlugin";
+import LinkActionPlugin from "../LinkActionPlugin";
 import URI from 'urijs';
 import i18n from "../../i18n";
 
-// See mime types discussion for 3d-tiles here and there
-// https://github.com/opengeospatial/ogcapi-3d-geovolumes/issues/13
-const OGC3DTILES_SUPPORTED_TYPES = [
-  // 'application/json',
-  'application/3dtiles+json',
-];
-
-export default class CesiumSandcastle extends AssetActionPlugin {
+export default class Cesium extends LinkActionPlugin {
 
   get show() {
-    return this.component.isBrowserProtocol && OGC3DTILES_SUPPORTED_TYPES.includes(this.asset.type);
+    return this.link.rel === '3d-tiles';
   }
 
-  get uri_CesiumSandcastle() {
-    let uri = new URI("https://sandcastle.cesium.com/index.html");
-    const tileset_url = this.component.href;
+  get uri() {
+    // https://sandcastle.cesium.com/standalone.html vs https://sandcastle.cesium.com/index.html
+    let uri = new URI("https://sandcastle.cesium.com/standalone.html");
+    const tileset_url = this.link.href;
     const code_payload = {
       html: `
         <style> @import url(../templates/bucket.css); </style>
@@ -42,14 +36,8 @@ export default class CesiumSandcastle extends AssetActionPlugin {
     return uri;
   }
 
-  get uri() {
-    let uri = new URI("https://viewer.geofox.ai/");
-    uri.addQuery('tileset', encodeURI(this.component.href));
-    return uri;
-  }
-
   get text() {
-    return i18n.t('actions.openIn', {service: 'Geofox.ai'});
+    return i18n.t('actions.openIn', {service: 'Cesium Sandcastle'});
   }
 
 }
